HTMLDocument element added to the constructor function of the library. This allows for _(document). However it would be nice to extend this in a more suitable way to allow for all HTML elements, ie body too.
Started to add Haskell function definitions for clarity purposes of what the parameters of functions are.

Added an error to be thrown if trying to encapsulate an object the library cannot support.

css now supports multiple setters and a single getter on the same function. Used like elem.css("left") and elem.css({"left" : "10", ...})

The On and Clear functions are in need of a cleanup and rethink. Particularly the Clear function for removing events.